pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aicoder"
version = "0.1.0"
description = "Automated PHP documentation tool with PHPDoc and section comments"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "typer >=0.9.0",
    "rich >=13.0.0",
    "python-dotenv >=1.0.0",
    "requests >=2.31.0",
    "openai>=1.60.1",
    "pyyaml>=6.0.2",
]

[project.scripts]
aicoder = "aicoder.cli.main:main"
patch = "aicoder.cli.patch:main"
test-patcher-v3 = "aicoder.cli.test_patcher_v3:main"

